tests/wpt/tests/css/css-text/text-align/text-align-last-interpolation.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>text-align-last interpolation</title>
    <link rel="author" title="Kiet Ho" href="mailto:tho22@apple.com">
    <link rel="help" href="https://drafts.csswg.org/css-text/#text-align-last-property">
    <meta name="assert" content="text-align-last supports discrete animation">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/css/support/interpolation-testcommon.js"></script>
  </head>
  <body>
    <script>
      test_no_interpolation({
        property: 'text-align-last',
        from: 'auto',
        to: 'start'
      });

      test_no_interpolation({
        property: 'text-align-last',
        from: 'justify',
        to: 'center'
      });

      test_no_interpolation({
        property: 'text-align-last',
        from: 'left',
        to: 'right'
      });

      test_no_interpolation({
        property: 'text-align-last',
        from: 'end',
        to: 'match-parent'
      });
    </script>
  </body>
</html>
