third_party/blink/web_tests/external/wpt/css/css-text/text-align/text-align-last-interpolation-expected.txt

This is a testharness.js-based test.
Found 112 tests; 75 PASS, 37 FAIL, 0 TIMEOUT, 0 NOTRUN.
FAIL CSS Transitions: property <text-align-last> from [auto] to [start] at (-0.3) should be [start] assert_equals: expected "start " but got "auto "
FAIL CSS Transitions: property <text-align-last> from [auto] to [start] at (0) should be [start] assert_equals: expected "start " but got "auto "
FAIL CSS Transitions: property <text-align-last> from [auto] to [start] at (0.3) should be [start] assert_equals: expected "start " but got "auto "
PASS CSS Transitions: property <text-align-last> from [auto] to [start] at (0.5) should be [start]
PASS CSS Transitions: property <text-align-last> from [auto] to [start] at (0.6) should be [start]
PASS CSS Transitions: property <text-align-last> from [auto] to [start] at (1) should be [start]
PASS CSS Transitions: property <text-align-last> from [auto] to [start] at (1.5) should be [start]
PASS CSS Transitions with transition: all: property <text-align-last> from [auto] to [start] at (-0.3) should be [start]
PASS CSS Transitions with transition: all: property <text-align-last> from [auto] to [start] at (0) should be [start]
PASS CSS Transitions with transition: all: property <text-align-last> from [auto] to [start] at (0.3) should be [start]
PASS CSS Transitions with transition: all: property <text-align-last> from [auto] to [start] at (0.5) should be [start]
PASS CSS Transitions with transition: all: property <text-align-last> from [auto] to [start] at (0.6) should be [start]
PASS CSS Transitions with transition: all: property <text-align-last> from [auto] to [start] at (1) should be [start]
PASS CSS Transitions with transition: all: property <text-align-last> from [auto] to [start] at (1.5) should be [start]
PASS CSS Animations: property <text-align-last> from [auto] to [start] at (-0.3) should be [auto]
PASS CSS Animations: property <text-align-last> from [auto] to [start] at (0) should be [auto]
PASS CSS Animations: property <text-align-last> from [auto] to [start] at (0.3) should be [auto]
PASS CSS Animations: property <text-align-last> from [auto] to [start] at (0.5) should be [start]
PASS CSS Animations: property <text-align-last> from [auto] to [start] at (0.6) should be [start]
PASS CSS Animations: property <text-align-last> from [auto] to [start] at (1) should be [start]
PASS CSS Animations: property <text-align-last> from [auto] to [start] at (1.5) should be [start]
PASS Web Animations: property <text-align-last> from [auto] to [start] at (-0.3) should be [auto]
PASS Web Animations: property <text-align-last> from [auto] to [start] at (0) should be [auto]
PASS Web Animations: property <text-align-last> from [auto] to [start] at (0.3) should be [auto]
PASS Web Animations: property <text-align-last> from [auto] to [start] at (0.5) should be [start]
PASS Web Animations: property <text-align-last> from [auto] to [start] at (0.6) should be [start]
PASS Web Animations: property <text-align-last> from [auto] to [start] at (1) should be [start]
PASS Web Animations: property <text-align-last> from [auto] to [start] at (1.5) should be [start]
FAIL CSS Transitions: property <text-align-last> from [justify] to [center] at (-0.3) should be [center] assert_equals: expected "center " but got "justify "
FAIL CSS Transitions: property <text-align-last> from [justify] to [center] at (0) should be [center] assert_equals: expected "center " but got "justify "
FAIL CSS Transitions: property <text-align-last> from [justify] to [center] at (0.3) should be [center] assert_equals: expected "center " but got "justify "
PASS CSS Transitions: property <text-align-last> from [justify] to [center] at (0.5) should be [center]
PASS CSS Transitions: property <text-align-last> from [justify] to [center] at (0.6) should be [center]
PASS CSS Transitions: property <text-align-last> from [justify] to [center] at (1) should be [center]
PASS CSS Transitions: property <text-align-last> from [justify] to [center] at (1.5) should be [center]
PASS CSS Transitions with transition: all: property <text-align-last> from [justify] to [center] at (-0.3) should be [center]
PASS CSS Transitions with transition: all: property <text-align-last> from [justify] to [center] at (0) should be [center]
PASS CSS Transitions with transition: all: property <text-align-last> from [justify] to [center] at (0.3) should be [center]
PASS CSS Transitions with transition: all: property <text-align-last> from [justify] to [center] at (0.5) should be [center]
PASS CSS Transitions with transition: all: property <text-align-last> from [justify] to [center] at (0.6) should be [center]
PASS CSS Transitions with transition: all: property <text-align-last> from [justify] to [center] at (1) should be [center]
PASS CSS Transitions with transition: all: property <text-align-last> from [justify] to [center] at (1.5) should be [center]
PASS CSS Animations: property <text-align-last> from [justify] to [center] at (-0.3) should be [justify]
PASS CSS Animations: property <text-align-last> from [justify] to [center] at (0) should be [justify]
PASS CSS Animations: property <text-align-last> from [justify] to [center] at (0.3) should be [justify]
PASS CSS Animations: property <text-align-last> from [justify] to [center] at (0.5) should be [center]
PASS CSS Animations: property <text-align-last> from [justify] to [center] at (0.6) should be [center]
PASS CSS Animations: property <text-align-last> from [justify] to [center] at (1) should be [center]
PASS CSS Animations: property <text-align-last> from [justify] to [center] at (1.5) should be [center]
PASS Web Animations: property <text-align-last> from [justify] to [center] at (-0.3) should be [justify]
PASS Web Animations: property <text-align-last> from [justify] to [center] at (0) should be [justify]
PASS Web Animations: property <text-align-last> from [justify] to [center] at (0.3) should be [justify]
PASS Web Animations: property <text-align-last> from [justify] to [center] at (0.5) should be [center]
PASS Web Animations: property <text-align-last> from [justify] to [center] at (0.6) should be [center]
PASS Web Animations: property <text-align-last> from [justify] to [center] at (1) should be [center]
PASS Web Animations: property <text-align-last> from [justify] to [center] at (1.5) should be [center]
FAIL CSS Transitions: property <text-align-last> from [left] to [right] at (-0.3) should be [right] assert_equals: expected "right " but got "left "
FAIL CSS Transitions: property <text-align-last> from [left] to [right] at (0) should be [right] assert_equals: expected "right " but got "left "
FAIL CSS Transitions: property <text-align-last> from [left] to [right] at (0.3) should be [right] assert_equals: expected "right " but got "left "
PASS CSS Transitions: property <text-align-last> from [left] to [right] at (0.5) should be [right]
PASS CSS Transitions: property <text-align-last> from [left] to [right] at (0.6) should be [right]
PASS CSS Transitions: property <text-align-last> from [left] to [right] at (1) should be [right]
PASS CSS Transitions: property <text-align-last> from [left] to [right] at (1.5) should be [right]
PASS CSS Transitions with transition: all: property <text-align-last> from [left] to [right] at (-0.3) should be [right]
PASS CSS Transitions with transition: all: property <text-align-last> from [left] to [right] at (0) should be [right]
PASS CSS Transitions with transition: all: property <text-align-last> from [left] to [right] at (0.3) should be [right]
PASS CSS Transitions with transition: all: property <text-align-last> from [left] to [right] at (0.5) should be [right]
PASS CSS Transitions with transition: all: property <text-align-last> from [left] to [right] at (0.6) should be [right]
PASS CSS Transitions with transition: all: property <text-align-last> from [left] to [right] at (1) should be [right]
PASS CSS Transitions with transition: all: property <text-align-last> from [left] to [right] at (1.5) should be [right]
PASS CSS Animations: property <text-align-last> from [left] to [right] at (-0.3) should be [left]
PASS CSS Animations: property <text-align-last> from [left] to [right] at (0) should be [left]
PASS CSS Animations: property <text-align-last> from [left] to [right] at (0.3) should be [left]
PASS CSS Animations: property <text-align-last> from [left] to [right] at (0.5) should be [right]
PASS CSS Animations: property <text-align-last> from [left] to [right] at (0.6) should be [right]
PASS CSS Animations: property <text-align-last> from [left] to [right] at (1) should be [right]
PASS CSS Animations: property <text-align-last> from [left] to [right] at (1.5) should be [right]
PASS Web Animations: property <text-align-last> from [left] to [right] at (-0.3) should be [left]
PASS Web Animations: property <text-align-last> from [left] to [right] at (0) should be [left]
PASS Web Animations: property <text-align-last> from [left] to [right] at (0.3) should be [left]
PASS Web Animations: property <text-align-last> from [left] to [right] at (0.5) should be [right]
PASS Web Animations: property <text-align-last> from [left] to [right] at (0.6) should be [right]
PASS Web Animations: property <text-align-last> from [left] to [right] at (1) should be [right]
PASS Web Animations: property <text-align-last> from [left] to [right] at (1.5) should be [right]
FAIL CSS Transitions: property <text-align-last> from [end] to [match-parent] at (-0.3) should be [match-parent] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions: property <text-align-last> from [end] to [match-parent] at (0) should be [match-parent] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions: property <text-align-last> from [end] to [match-parent] at (0.3) should be [match-parent] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions: property <text-align-last> from [end] to [match-parent] at (0.5) should be [match-parent] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions: property <text-align-last> from [end] to [match-parent] at (0.6) should be [match-parent] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions: property <text-align-last> from [end] to [match-parent] at (1) should be [match-parent] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions: property <text-align-last> from [end] to [match-parent] at (1.5) should be [match-parent] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <text-align-last> from [end] to [match-parent] at (-0.3) should be [match-parent] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <text-align-last> from [end] to [match-parent] at (0) should be [match-parent] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <text-align-last> from [end] to [match-parent] at (0.3) should be [match-parent] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <text-align-last> from [end] to [match-parent] at (0.5) should be [match-parent] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <text-align-last> from [end] to [match-parent] at (0.6) should be [match-parent] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <text-align-last> from [end] to [match-parent] at (1) should be [match-parent] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <text-align-last> from [end] to [match-parent] at (1.5) should be [match-parent] assert_true: 'to' value should be supported expected true got false
FAIL CSS Animations: property <text-align-last> from [end] to [match-parent] at (-0.3) should be [end] assert_true: 'to' value should be supported expected true got false
FAIL CSS Animations: property <text-align-last> from [end] to [match-parent] at (0) should be [end] assert_true: 'to' value should be supported expected true got false
FAIL CSS Animations: property <text-align-last> from [end] to [match-parent] at (0.3) should be [end] assert_true: 'to' value should be supported expected true got false
FAIL CSS Animations: property <text-align-last> from [end] to [match-parent] at (0.5) should be [match-parent] assert_true: 'to' value should be supported expected true got false
FAIL CSS Animations: property <text-align-last> from [end] to [match-parent] at (0.6) should be [match-parent] assert_true: 'to' value should be supported expected true got false
FAIL CSS Animations: property <text-align-last> from [end] to [match-parent] at (1) should be [match-parent] assert_true: 'to' value should be supported expected true got false
FAIL CSS Animations: property <text-align-last> from [end] to [match-parent] at (1.5) should be [match-parent] assert_true: 'to' value should be supported expected true got false
FAIL Web Animations: property <text-align-last> from [end] to [match-parent] at (-0.3) should be [end] assert_true: 'to' value should be supported expected true got false
FAIL Web Animations: property <text-align-last> from [end] to [match-parent] at (0) should be [end] assert_true: 'to' value should be supported expected true got false
FAIL Web Animations: property <text-align-last> from [end] to [match-parent] at (0.3) should be [end] assert_true: 'to' value should be supported expected true got false
FAIL Web Animations: property <text-align-last> from [end] to [match-parent] at (0.5) should be [match-parent] assert_true: 'to' value should be supported expected true got false
FAIL Web Animations: property <text-align-last> from [end] to [match-parent] at (0.6) should be [match-parent] assert_true: 'to' value should be supported expected true got false
FAIL Web Animations: property <text-align-last> from [end] to [match-parent] at (1) should be [match-parent] assert_true: 'to' value should be supported expected true got false
FAIL Web Animations: property <text-align-last> from [end] to [match-parent] at (1.5) should be [match-parent] assert_true: 'to' value should be supported expected true got false
Harness: the test ran to completion.
